+++ apps/web/src/features/address/ui/index.ts
@@ -2,5 +2,4 @@ export { AddressCard } from './AddressCard';
 export { AddressEmpty } from './AddressEmpty';
 export { AddressForm } from './AddressForm';
 export { KakaoPostcode } from './KakaoPostcode';
-export type { AddressFormMode, AddressFormValues } from './types';
 export type { AddressFormProps } from './AddressForm';
